major rework with adding files and revising code
