+++ src/ui/mod.rs
@@ -1,9 +1,6 @@
-pub mod components;
 mod home;
 mod popups;
-// pub use components;
-pub use home::centered_rect;
-pub use home::ui;
+pub use home::{centered_rect, ui};
 pub use popups::{
     add_request_popup, editing_popup, exiting_popup, render_header_popup, render_params_popup,
 };
