+++ src/ui_components/mod.rs
@@ -1,0 +1,3 @@
+pub mod details;
+pub mod footer;
+pub mod groups;
